Separate notebooks_test_utils functions from notebooks_test.py
